requirements.txt
astroid==1.6.1
backports.functools-lru-cache==1.5
configparser==3.5.0
Django==1.11.10
django-bootstrap-ui==0.5.1
django-bootstrap3==9.1.0
django-tag-parser==2.1
dominate==2.1.17
enum34==1.1.6
futures==3.2.0
isort==4.3.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pytz==2018.3
PyYAML==3.12
singledispatch==3.4.0.3
six==1.11.0
termcolor==1.1.0
wrapt==1.10.11
